// client/src/app/core/models/title.model.ts
export interface TitleItemResponse {
    id: string;
    tipo: string;
    archivo: string;
    archivo_numero: number;
    archivo_completo: string
    titulo: string
    titulo_num: number
    productor_id: string
    descripcion: string
    rfc: string
    curp: string
    expedidoa2: string
    expedidoa_apellidos: string
    expedidos_nombres: string
    criador: string
    criador_de: string
    asiento_produccion: string
    municipio: string
    avecinado: string
    transferidoa: string
    status: string
    puesto_firma: string
    nombre_firma: string
    fecha_expedicion: string
    observaciones: string
    estatus_captura_datos: string
    usuario_captura_datos: string
    fecha_captura_datos: string
    estatus_captura_imagen: string
    usuario_captura_imagen: string
    fecha_captura_imagen: string
    recuperada: string
    senal_lista: string
    marca_lista: string

}